requirements.txt
Pillow==8.3.1
requests==2.25.1